Factor known product ID prefixes out to a list.
This will make adding more prefixes easier in the future.

We also add the Wii prefix to prevent it from being duplicated in case
the user types it by mistake. We enforce it to 4 characters to make
our code below easier: we can get the first 4 chars of the user input
and search in the list, instead of making specific checks depending on
the system.