Detect system from product ID.
Product IDs have prefixes unique per system, so we can easily guess
the system from the set of known prefixes.

This removes the need for the system configuration variable, and
allows to enter games from different systems in the same session.

As a small drawback, you have to type the full code the first time,
but after that the last prefix will be offered as the default so I
don't think it's a big problem.