Ask users to enter a system if one was not identified.
This allows the script to still be useful to enter games for which we
haven't coded a system yet (or which cannot be easily guessed, like PC
games).

It has drawbacks: it only works if the game from the unknown system is
entered in first place, once we guess one system we stick to it. Also,
you have to type it every time.